Don't re-prompt for email if account already exists
Extracting CertMagic out of Caddy was a new and complicated surgery,
and although successful, it wasn't without its side-effects. One of
them was we slightly altered some of the subtle logic related to
getting an email address in order to know which account to use.

If an ACME account with an empty email already exists, and no Email is
available on the config, that account should be used without prompting
the user over and over. If no account exists and no email is available,
then yes a prompt is warranted. And of course if an email is set or an
account with a non-empty email exists, that should be used without
prompts (this was already the case).

This commit fixes to align with that expectation, so that accounts with
empty emails can be reused, if present.

Empty emails should only be used in dev or testing environments;
omitting them in production is bad practice.