Background cert operations; ManageSync() and ManageAsync()
Split Manage() into ManageSync() and ManageAsync().

In accordance with developing best practices, ACME operations should be
allowed to happen in the background and not block server startup in
non-interactive environments.

We also no longer return an error during batch cert renewals, because
we always treat it as a background operation. (The ManageSync() method
can perform foreground renewal if that is desired.)