Treat on-demand cert ops as interactive to avoid long-lived retries
Although the server is not currently interactive, TLS handshakes happen
in the foreground, so we pretend as if it's interactive to return
quickly if there's an error.

TODO: This could potentially lead to a prompt while the server is
running; probably better to detect interactive terminal instead, and
make the boolean argument more about whether to enable retries.