Retry DNS query with TCP if UDP fails (#81)
* Minor improvement to DNS request handling

Sometimes incoming udp traffic on port 53 is blocked to
prevent DDoS attacks. In those cases only TCP will work
for DNS request as the UDP request will time out. And as
a result the DNS challenge will fail, while the server is
trying to verify if the challenge was propageted through
the NS.

Now instead of returning immidently, if a timeout with UDP was
received, the request will be tried again using TCP.

* Formatting and comment

Co-authored-by: Georg Friedrich <g.friedrich@sonnenwagen.org>
Co-authored-by: Matthew Holt <mholt@users.noreply.github.com>