Don't maintain on-demand certs in background
On-demand certs are managed at handshake-time. Doing so in the background was
a temporary holdover until on-demand maintenance improved, which it since has.
Since background maintenance did not consult the "ask" endpoint or decision func,
it would sometimes renew certificates that were not desirable to renew.

See https://caddy.community/t/clean-up-caddy-certificates/11429/11?u=matt