Fixed order of certificate loading so that private keys are loaded first
The order of storing the certificates was previously changed so that the
private key would be stored first. For anyone who is creating storage
hooks which push the certificate & key to a 3rd party service (like a
CDN), the certificates are generally refused if uploaded before the
private key.

Loading can trigger the same issue if (for any reason), the certificate
& key have been deleted on the 3rd party service.

I'll admit that this stretches what the storage system should be doing,
but with this trivial change it is really easy to create a reliable
link to a CDN.