Always call checkDNSPropagation in DNS01Solver
Calling checkAuthoritativeNss is wrong as it's not inter-changeable with
checkDNSPropagation.

Though IIUC it's not necessary to follow CNAME when in Wait(), with or
without OverrideDomain, let's wait until the override domain gets some
usage to change this. The reason that following CNAME is not necessary
is that CNAME cannot co-exist with other DNS records, if we succeed in
setting a TXT record on that domain, it cannot have a CNAME record.

@IndeedNotJames