Account for NotAfter imprecision in timestamp comparisons
The NotAfter field of an X.509 certificate has a one second resolution.
Because of the imprecise definition of NotAfter, the certificate is
valid for the entire second of a NotAfter timestamp. This effectively
means the certificate expires at the next second after the NotAfter
timestamp.

Add the expiresAt function to account for this extra validity time when
performing time comparisons against other timestamps. Timestamp like the
NextUpdate of an ocsp response do not have this imprecision and do not
need to be adjusted when calculating.

Fixes #197